Move HLD from graph to tree
