+++ test/tree/heavy_light_decomposition/hld.test.cpp
@@ -1,6 +1,6 @@
 #define PROBLEM "https://onlinejudge.u-aizu.ac.jp/courses/library/5/GRL/5/GRL_5_E"
 
-#include "../../../lib/graph/heavy_light_decomposition.cpp"
+#include "../../../lib/tree/heavy_light_decomposition.cpp"
 #include "../../../lib/structure/lazy_segment_tree.cpp"
 
 int main()
@@ -16,7 +16,6 @@ int main()
     }
 
     HLDecomposition hld(g);
-    hld.build();
 
     LazySegmentTree<ll> segt(
             n,
